requirements.txt
absl-py==2.1.0
affine==2.3.1
appnope==0.1.3
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
brewer2mpl==1.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
clarabel==0.7.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.2
contextily==1.2.0
contourpy==1.0.6
cvxpy==1.4.3
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
deepdiff==7.0.1
dotenv
ecos==2.0.13
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.16.1
#Fiona==1.8.22 installed using conda due to GDAL stuff
fonttools==4.38.0
fsspec==2024.10.0
gcloud==0.18.3
geographiclib==2.0
geopandas==0.12.1
geopy==2.3.0
ggplot==0.11.5
googleapis-common-protos==1.58.0
httplib2==0.21.0
huggingface-hub==0.26.2
idna==3.4
importlib_metadata==7.1.0
ipykernel==6.19.2
ipython==8.7.0
jax==0.4.28
#jax-metal==0.0.7 removed due to only existing for MACOS
jaxlib==0.4.28
jedi==0.18.2
Jinja2==3.1.4
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
MarkupSafe==3.0.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mercantile==1.2.1
mizani==0.8.1
#ml-dtypes==0.2.0 #installed 0.5.0 instead
mpmath==1.3.0
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.2.1
#numpy==1.23.5 installed 2.2.1 instead


oauth2client==4.1.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
osqp==0.6.5
packaging==22.0
palettable==3.3.0
pandapower==2.14.6
pandas==2.2.2
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.1.1
platformdirs==2.6.0
plotnine==0.10.1
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.12.0
Pygments==2.13.0
pyparsing==3.0.9
#pyproj==3.4.0 installed 3.7.0 instead
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0.2
#pyzmq==24.0.1 installed 26.2.0 instead


qdldl==0.1.7.post2
#rasterio==1.3.4 installed 1.4.3 instead


regex==2024.11.6
requests==2.28.1
rsa==4.9
safetensors==0.4.5


#scikit-learn==1.2.0 installed 1.6.0 instead
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.12.2
setuptools==62.3.2
# Shapely==1.8.5.post1 installed 2.2.1 instead
six==1.16.0
# scikit-learn==1.2.0 installed 1.6.0 instead
snuggs==1.4.7
stack-data==0.6.2
#statsmodels==0.13.5 installed 0.14.4 instead
sympy==1.13.3
threadpoolctl==3.1.0
tokenizers==0.20.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.2
tqdm==4.66.4
traitlets==5.7.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.13
wcwidth==0.2.5
wheel==0.43.0
xarray==2024.3.0
xyzservices==2022.9.0
zipp==3.18.2


